lambda/requirements.txt
boto3==1.34.11
python-dateutil==2.8.2
